pyproject.toml
[project]
name = "langhchain"
version = "0.1.0"
description = "LangChain project"
authors = [
    {name = "Sebastian Zapata"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain (>=1.2.15,<2.0.0)",
    "langchain-openai (>=1.2.0,<2.0.0)",
    "langchain-core (>=1.3.2,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langgraph (>=1.1.9,<2.0.0)",
    "langchain-postgres (>=0.0.17,<0.0.18)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "langchain-text-splitters (>=1.1.2,<2.0.0)",
    "langchain-ollama (>=1.1.0,<2.0.0)",
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pypdf (>=6.10.2,<7.0.0)",
    "langchain-huggingface (>=0.0.3,<1.0.0)",
    "transformers (>=4.40.0,<5.0.0)",
    "torch (>=2.0.0,<3.0.0)",
    "sentence-transformers (>=2.7.0,<3.0.0)",
    "accelerate (>=0.30.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
